output and input space have same type

+++ stencil_code/stencil_kernel.py
@@ -164,7 +164,7 @@ class OclStencilFunction(ConcreteSpecializedFunction):
                 and isinstance(args[0], hmarray):
             output = empty_like(args[0])
         elif self.lsf.num_convolutions > 1:
-            output = np.zeros((self.lsf.num_convolutions,) + args[0].shape)
+            output = np.zeros((self.lsf.num_convolutions,) + args[0].shape).astype(args[0].dtype)
         else:
             output = np.zeros_like(args[0])
         # self.kernel.argtypes = tuple(
